fix: don't write package-lock.json
npm v7 introduced lockfile version 2. When running earlier versions (eg:
on a github actions runner) npm will warn with:

```
npm WARN read-shrinkwrap This version of npm is compatible with lockfileVersion@1, but package-lock.json was generated for lockfileVersion@2. I'll try to do my best with it!
```

npm v6 then overwrites package-lock.json with a version 1 lockfile.
This causes the pre-commit hooks to fail, example:
https://github.com/tekumara/asu/runs/2212404614?check_suite_focus=true#step:6:106