fix: add wheel
Add wheel to remove the warning on dependencies:

```
Using legacy 'setup.py install' for tesserocr, since package 'wheel' is not installed.
Using legacy 'setup.py install' for pdfplumber, since package 'wheel' is not installed.
Using legacy 'setup.py install' for pytesseract, since package 'wheel' is not installed.
Using legacy 'setup.py install' for antlr4-python3-runtime, since package 'wheel' is not installed.
```

wheel was removed in 74b8e90 and specified in the `build-system` of _pyproject.toml_. However it turns out that only applies to this project and not its dependencies, which may not specify a `build-system` and so expect wheel to be there before they are installed.